初始化支付接口完成

+++ src/main/java/com/cn/school/dto/info/bo/AddOrderBO.java
@@ -79,7 +79,7 @@ public class AddOrderBO extends UserContextViewForm{
     /**
      * 总金额
      */
-    private Long totalFee;
+    private Integer totalFee;
     /**
      *电子发票
      */
