+++ src/main/java/com/cn/school/entity/DSOrder.java
@@ -73,7 +73,7 @@ public class DSOrder {
     /**
      * 总金额
      */
-    private Long totalFee;
+    private Integer totalFee;
     /**
      *电子发票
      */
